// frontend/e-commerce-service/src/app/generated/model/orderItemRequestDTO.ts
/**
 * e-commerce-service
 * API documentation for the e-commerce-service
 *
 * OpenAPI spec version: v1
 * Contact: placeholder@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface OrderItemRequestDTO { 
    orderID: string;
    productID: string;
    quantity: number;
    price: number;
}